refactor: move ZWO and UVC files into Services/Zwo/ and Services/Uvc/ subdirectories
- Move ZWO-related files (ZWOASICameraInterop, ZWOCameraDetect,
  ZwoFrameSource, ZWOLiveStreamService, IZwoFrameSource) to Services/Zwo/
- Move UVC-related files (UvcFrameSource, IUvcFrameSource, LibUvcInterop,
  IokitHelper) to Services/Uvc/
- Update namespaces from CollimationCircles.Services to
  CollimationCircles.Services.Zwo and CollimationCircles.Services.Uvc
- Add using directives in all referencing files
