+++ CollimationCircles/Services/Uvc/UvcFrameSource.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 
-namespace CollimationCircles.Services
+namespace CollimationCircles.Services.Uvc
 {
     /// <summary>
     /// Direct UVC camera frame source using the libuvc library
